Fix: update Permision dropdown
